refactor: moving files to new recipes folder
