+++ src/main/java/com/example/MG_RestaurantManager20/recipe/adapters/web/RecipeController.java
@@ -1,7 +1,7 @@
-package com.example.MG_RestaurantManager20.product.adapters.web;
+package com.example.MG_RestaurantManager20.recipe.adapters.web;
 
-import com.example.MG_RestaurantManager20.product.domain.Recipe;
-import com.example.MG_RestaurantManager20.product.service.ports.RecipeService;
+import com.example.MG_RestaurantManager20.recipe.domain.Recipe;
+import com.example.MG_RestaurantManager20.recipe.service.ports.RecipeService;
 import com.example.MG_RestaurantManager20.product.struct.ProductStructure;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
