+++ src/main/java/com/example/MG_RestaurantManager20/recipe/service/ports/RecipeService.java
@@ -1,7 +1,7 @@
-package com.example.MG_RestaurantManager20.product.service.ports;
+package com.example.MG_RestaurantManager20.recipe.service.ports;
 
-import com.example.MG_RestaurantManager20.product.adapters.database.RecipeRepository;
-import com.example.MG_RestaurantManager20.product.domain.Recipe;
+import com.example.MG_RestaurantManager20.recipe.adapters.database.RecipeRepository;
+import com.example.MG_RestaurantManager20.recipe.domain.Recipe;
 import com.example.MG_RestaurantManager20.product.struct.ProductStructure;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
